src/cmsplugin_slick_gallery/translation.py
from modeltranslation.translator import translator, TranslationOptions
from .models import GalleryImage

class GalleryImageTranslationOptions(TranslationOptions):
    fields = ('text',)

translator.register(GalleryImage, GalleryImageTranslationOptions)