requirements.txt
Django >= 1.11
django-filer >= 1.2.0
django-cms >= 3.4
djangocms-text-ckeditor